+++ Cargo.toml
@@ -1,11 +1,16 @@
-[package]
-name = "anchor"
-version = "0.1.0"
-edition = "2021"
-authors = ["Sigma Prime <contact@sigmaprime.io>"]
-rust-version = "1.80.0"
+[workspace]
+# Extra tooling projects will be added.
+members = [
+    "anchor",
+    "anchor/client",
+]
+resolver = "2"
 
-[dependencies]
+[workspace.package]
+edition = "2021"
+
+[workspace.dependencies]
+client = { path = "anchor/client" }
 task_executor = { git = "https://github.com/sigp/lighthouse", branch = "anchor", default-features = false,  features = ["tracing"] }
 sensitive_url = { git = "https://github.com/sigp/lighthouse", branch = "anchor"}
 async-channel = "1.9"
@@ -20,5 +25,8 @@ tracing = "0.1.40"
 git-version = "0.3.9"
 target_info = "0.1.0"
 
-[dev-dependencies]
-regex = "1.10.6"
+[profile.maxperf]
+inherits = "release"
+lto = "fat"
+codegen-units = 1
+incremental = false
